For bug 2237:
Windows installers now create a registry key in the HKEY_LOCAL_MACHINE
hive to allow us to store Ice installation specific data in a well-known
location (e.g. indicate where Ice is installed). The key is of the form
HKEY_LOCAL_MACHINE\SOFTWARE\ZeroC\[ProductName]. Currently the only data
stored under this key is the installation directory. The directory is a
string value with the name "InstallDir".