Fixed bug in TestUtil.py that caused line numbers to be omitted from the stack trace under Mono when the a test failed.
Ported servant locator exceptions.

Fixed bug that caused the "unknown" member of unknown exception to not be initialized
in some cases, and initialized to the wrong value in other cases.

Added ice_name() to mapping for exceptions--without it, there was no way to get at
the Slice type id of an exception, which was needed to set the "unknown" member
correctly.