Fix adapter list update problem on replica icegridregistries.
When changing adapters in an icebox server (e.g. adding an adapter),
a replicated registry still uses the previous list of
adapter ids - which means, an icegridnode starting a service fails
to do so in case it's using a replicated registry. As it's common
to have one master registry and many replicated ones, this gets
worse the more replicated registries exist.