Fix windows IceSSL/configuration failure
The test failed in Windows because openssl emit a warning when
running openssl command related to the openssl config file being
missed.

We previously deal with this by only reading the first line of the
output but that is a be fragile as WARNING can be in the first line
as show in the test failures.