Added nullable type support to the Java mapping
Added support for nullable and non-null types in Java through annotations provided by Checker Framework. These have replaced the old java.util.Optional wrappers we used to use for this. It is part of #611.

Also added a type-checker to the build system that enforces nullness during compile time. Right now the checker is disabled since there's alot of warnings (which the build system treats as errors and fails) with the mapping, that I'm still in the process of fixing. After fixing them, I'll re-enable the checker.

This commit adds a dependency on the Checker Framework library.