Updated Ice to version 4.0.0-alpha.0 (#690)
Updated the version of Ice to '4.0.0-alpha.0'

Also reworked alot of versioning and version format logic:

-Removed our old methodology of treating patch versions '>50'=alpha versions and '>60'=beta versions.
-Now for alpha releases we define 'ICE_ALPHA_VERSION' to be the release number, and same for beta releases but we instead define 'ICE_BETA_VERSION'.
-Also switched alot of the version formats we used throughout Ice (and related features) to use semver format: 'www.semver.org'.

Temporarily disabled building and testing of the Python mapping. It should be re-enabled once it's been re-written. Because of this I've switched MATLAB to use Java for it's testing servers (since it was previously using Python). Have also permanently disabled testing of the cpp98 directory.

Cleaned up the checked in generated code. Removed generated version strings and the version mismatch checks we used in C++.