Added support for new integer types to js compiler.
This 'support' doesn't extend to unsigned longs. Right now I've mapped them to our own
Ice.Long type, which doesn't support the full unsigned long range and is a signed type.
In a future PR Ice.Long should be replaced with the new BigInt type. Support for
unsigned longs should be added when this is done.