Adds support for new integer types to the Slice compiler. (#756)
This pull request adds support for the following unsigned, and variable length integers in the Slice compiler:
ushort, uint, ulong, varint, varuint, varlong, varulong

It also provides preliminary support and a mapping for these types in all languages except Javascript and Java, due to the lack of a trivial mapping for ulong in these languages. Note that each new type still needs to be implemented in each prospective language mapping.