Added back check for x86 for multilib platforms #773
The check is required when building RPM packages for CentOS/RHEL.

If building on a x86 platform which doesn't support multilib, either
"make PLATFORMS=x86" should be used or the linux_id variable should be set with
the linux ID of the distribution (make linux_id=<distribution>).

By default, the build system checks for ID_LIKE for the distribution type and
assumes a multilib platform if ID_LIKE is fedora, centos or rhel.