Add class cycle detection during unmarshaling
Add support for detection of class cycles during unmarshaling in
languages which do no have cyclic reference garbage collection:
C++, Swift, and Objective-C.

A MarshalException is thrown when a cycle is detected to prevent leaking
memory.

The property Ice.AcceptClassCycles can be used to change this behavior
and allow unmarshaling of class cycles.