Fix clang build (missing copy constructor)
In 15a9edabe things were moved around, so that clang/libc++
builds have the copy constructor of GroupNodeInfo declared, but
(unlike the assignment operator in NodeI.cpp#L101, which might
not be needed to spelled out anymore these days) not defined,
which results in broken builds at link time.

Tested with clang v11.0.1.