Do not use QueueExecutor for Android - Fix #3300
Ice no longer uses a  QueueExecutor on Android, as it was working around the restriction on making synchronous calls from
the main thread. Android explicitly prohibits IO operations on the main thread, which QueueExecutor was bypassing.

Additionally, QueueExecutor was prone to deadlocks since all IO operations were being executed on a single thread
(see #3300 for details).