Do not use QueueExecutor for Android - Fix #3300 (#3325)
Ice no longer uses a QueueExecutor on Android, as it was working around the restriction on making IO calls from the main thread. Android explicitly prohibits IO operations on the main thread, which QueueExecutor was bypassing.

Additionally, QueueExecutor was prone to deadlocks because all IO operations were executed on a single thread (see #3300 for details).

The Android test suite creates a Controller object. When the controller receives a runTestCase request, it creates a TestCase object, which can be used to start the server and client applications. These applications run in separate threads within the same process.

This setup means we are not testing whether async operations are safe to call from the main thread.